Add validation/execution priority order to samples
* For any "validation" property in a dataset or "policy" property in
an activity, give it a validationPriorityOrder or
executionPriorityOrder value of "OldestFirst", respectively, if one
is not present already .
* The new behavior of the service will be to always return the
priority order, regardless of its value when the user created it.
Previously, if validationPriorityOrder or executionPriorityOrder
were unspecified or the user specified "OldestFirst", the service
would not have serialized the value back. This gave the impression
that the property value was not being honored and was confusing
to users.